backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-decouple==3.8
Pillow==10.1.0
django-extensions==3.2.3 